scenario: ComparableAssertableValue (which ComparableAnswer now extends)
This just adds a new layer of inheritance for the same `Comparable*` base functionality, allowing for non-“answer” values to use the same comparison/assertion extension facilities.
